+++ src/main/java/wooteco/subway/dao/JdbcStationDao.java
@@ -10,7 +10,7 @@ import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
 
-import wooteco.subway.domain.Station;
+import wooteco.subway.domain.station.Station;
 
 @Repository
 public class JdbcStationDao implements StationDao {
