+++ src/main/java/wooteco/subway/dao/StationDao.java
@@ -2,7 +2,7 @@ package wooteco.subway.dao;
 
 import java.util.List;
 
-import wooteco.subway.domain.Station;
+import wooteco.subway.domain.station.Station;
 
 public interface StationDao {
     Station save(Station station);
