+++ src/main/java/wooteco/subway/domain/path/PathFindingStrategy.java
@@ -2,8 +2,8 @@ package wooteco.subway.domain.path;
 
 import java.util.List;
 
-import wooteco.subway.domain.Station;
 import wooteco.subway.domain.line.Lines;
+import wooteco.subway.domain.station.Station;
 
 public interface PathFindingStrategy {
     int getShortestDistance(Station source, Station target, Lines lines);
