+++ src/main/java/wooteco/subway/domain/station/StationRepository.java
@@ -1,0 +1,4 @@
+package wooteco.subway.domain.station;
+
+public interface StationRepository {
+}
