+++ src/main/java/wooteco/subway/service/SectionService.java
@@ -5,11 +5,11 @@ import org.springframework.transaction.annotation.Transactional;
 
 import wooteco.subway.dao.SectionDao;
 import wooteco.subway.dao.StationDao;
-import wooteco.subway.domain.Station;
 import wooteco.subway.domain.line.Line;
 import wooteco.subway.domain.line.LineRepository;
 import wooteco.subway.domain.section.Section;
 import wooteco.subway.domain.section.Sections;
+import wooteco.subway.domain.station.Station;
 import wooteco.subway.dto.service.request.SectionCreateRequest;
 import wooteco.subway.dto.service.request.SectionDeleteRequest;
