+++ src/test/java/wooteco/subway/dao/JdbcStationDaoTest.java
@@ -9,7 +9,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 
-import wooteco.subway.domain.Station;
+import wooteco.subway.domain.station.Station;
 
 @JdbcTest
 public class JdbcStationDaoTest {
