+++ src/test/java/wooteco/subway/domain/section/SectionsTest.java
@@ -7,7 +7,7 @@ import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import wooteco.subway.domain.Station;
+import wooteco.subway.domain.station.Station;
 
 public class SectionsTest {
     @DisplayName("구간을 추가한다.")
