+++ src/test/java/wooteco/subway/service/FakeStationDao.java
@@ -9,7 +9,7 @@ import java.util.Map;
 import org.springframework.util.ReflectionUtils;
 
 import wooteco.subway.dao.StationDao;
-import wooteco.subway.domain.Station;
+import wooteco.subway.domain.station.Station;
 
 public class FakeStationDao implements StationDao {
     private Long seq = 0L;
