+++ src/test/java/wooteco/subway/service/StationServiceTest.java
@@ -7,8 +7,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import wooteco.subway.dao.SectionDao;
-import wooteco.subway.domain.Station;
 import wooteco.subway.domain.section.Section;
+import wooteco.subway.domain.station.Station;
 import wooteco.subway.dto.service.StationDto;
 
 public class StationServiceTest {
