+++ src/test/java/wooteco/subway/service/repository/LineRepositoryWithDaoTest.java
@@ -11,10 +11,10 @@ import org.junit.jupiter.api.Test;
 import wooteco.subway.dao.LineDao;
 import wooteco.subway.dao.SectionDao;
 import wooteco.subway.dao.StationDao;
-import wooteco.subway.domain.Station;
 import wooteco.subway.domain.line.Line;
 import wooteco.subway.domain.section.Section;
 import wooteco.subway.domain.section.Sections;
+import wooteco.subway.domain.station.Station;
 import wooteco.subway.service.FakeLineDao;
 import wooteco.subway.service.FakeSectionDao;
 import wooteco.subway.service.FakeStationDao;
